services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
kiteconnect==4.2.0
python-dotenv==1.0.0
prometheus-client==0.19.0
psycopg2-binary==2.9.9
pika==1.3.2
